requirements.txt
# Flask and backend development
Flask>=2.0.0,<4.0.0
flask-wtf>=1.0.0
Werkzeug>=2.0.0,<4.0.0

# Google Gemini API (correct package name)
google-generativeai>=0.3.0

# Whisper and speech processing
edge-tts>=6.1.0
openai-whisper>=20231117

# Video and audio processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0
pydub>=0.25.1

# PDF processing
pdf2image>=1.16.0
Pillow>=8.0.0,<11.0.0

# Compatibility with Colab's numpy version
numpy>=1.21.0,<2.4.0

# OpenAI Whisper (use PyPI version for stability)
openai-whisper>=20231117

# System tools
tqdm
nest_asyncio>=1.5.6
SoundFile>=0.12.1
python-dotenv
PyPDF2>=2.0.0
ipython
pyngrok
